+++ src/index.js
@@ -1,6 +1,7 @@
 export { default as config } from './config.js';
 export { default as wiki } from './middleware.js';
 export { default as UttoriWiki } from './wiki.js';
+export { asyncHandler } from './wiki.js';
 export { default as wikiFlash } from './wiki-flash.js';
 export { default as AddQueryOutputToViewModel } from './plugins/add-query-output.js';
 export { default as DownloadRouter } from './plugins/add-download-route.js';